+++ utils/training.py
@@ -65,8 +65,6 @@ import torch.nn as nn
 import torch.nn.functional as F
 from tqdm import tqdm
 
-import sys
-sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 from configs.config import ViTConfig